+++ djpcms/permissions.py
@@ -1,8 +1,40 @@
-'''
-Default permission for inline editing
-'''
+'''Permission for inline editing'''
+from django.contrib.auth.backends import ModelBackend
 
+from djpcms.conf import settings
 from djpcms.utils import function_module
+
+
+
+def get_view_permission(obj):
+    return '%s_view' % obj._meta
+
+
+def has_permission(user, permission_codename, obj=None):
+    from djpcms.models import Page, BlockContent
+    if not obj:
+        return user.has_perm(permission_codename)
+    else:
+        anony = user.is_anonymous()
+        opts = obj._meta
+        
+        viewperm     = get_view_permission(obj) == permission_codename
+        changeperm   = opts.get_change_permission() == permission_codename
+        
+        # Do Page and BlockContent first
+        if isinstance(obj,Page):
+            if anony and obj.requires_login:
+                return False
+            if changeperm and obj.user == user and settings.DJPCMS_USER_CAN_EDIT_PAGES:
+                return True
+        elif isinstance(obj,BlockContent):
+            if changeperm and obj.page.user == user and settings.DJPCMS_USER_CAN_EDIT_PAGES:
+                return True
+            if viewperm and not anony and obj.for_not_authenticated:
+                return False
+        
+        return viewperm or sup_has_perm(user, permission_codename)
+        
 
 def _inline_editing(request, page, obj = None):
     '''
@@ -34,3 +66,13 @@ def inline_editing(request, page, obj = None):
             return False
     else:
         return False
+    
+    
+class Backend(ModelBackend):
+    '''Permission backend which complement the standard django backend.'''
+    supports_object_permissions = True
+    supports_anonymous_user = True
+
+    def has_perm(self, user, permission_codename, obj=None):
+        sup_has_perm = super(Backend,self).has_perm
+        return sup_has_perm(user, permission_codename)